refactor: 상세피드 구글맵 리팩터링
