+++ package.json
@@ -15,7 +15,7 @@
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.15.3",
     "@mui/material": "^5.15.2",
-    "@react-google-maps/api": "^2.19.2",
+    "@react-google-maps/api": "^2.19.3",
     "@react-pdf/renderer": "^3.1.15",
     "@react-spring/web": "^9",
     "@tanstack/react-query": "^5.15.0",
